backend/pyproject.toml
[tool.poetry]
name = "college-toolbox"
version = "0.1.0"
description = ""
authors = ["poggecci <alejandro.cruzado@upr.edu>"]
readme = "README.md"

packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.103.1"}
gunicorn = "^21.2.0"
sqlalchemy = "^2.0.20"
psycopg2-binary = "^2.9.7" # needed for postgres SQLAlchemy integration
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"
flake8 = "^6.1.0"
pytest = "^7.4.1"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
